refactor(auto_search): extract SeaDex lookup + cache into seadex_lookup.rs
services/auto_search/mod.rs was 2373 LoC; the SeaDex lookup / cache /
DB-persistence subsystem was the largest self-contained unit at ~550
LoC. Moves:

  mod.rs           1700  (was 2373) — auto-search pipeline + scoring
                          glue. Now imports the SeaDex entry points
                          from seadex_lookup via super-private re-
                          exports and a couple of `pub use`s for the
                          startup-warmup hooks.
  seadex_lookup.rs  711  SeaDexPayload struct, SEADEX_CACHE +
                          SEADEX_INFLIGHT statics, the seadex_cache_*
                          helpers, seadex_persist_to_db,
                          seadex_warm_cache_from_db,
                          prewarm_seadex_negative,
                          fetch_seadex_payload, seadex_gates +
                          is_seadex_match. Plus the four SeaDex tests
                          (`seadex_persist_round_trips_through_warm`,
                          `seadex_warm_drops_expired_persisted_rows`,
                          `seadex_error_cache_is_in_memory_only`,
                          `seadex_inflight_contains_tracks_registry`)
                          live alongside the code they exercise.

Module name is `seadex_lookup` rather than `seadex` to avoid
shadowing the `crate::services::seadex` import (the upstream
releases.moe client) that mod.rs already had in scope.

Verify: 1233 unit + 7 PR0 + 1 e2e tests all green, fmt + clippy
-D warnings clean.
